refactor: use only one Choose component
