+++ src/token/caption/Caption.jsx
@@ -1,15 +1,15 @@
 import React from 'react'
-import Choose from './Choose'
+import Choose from '../choose/Choose'
 import CaptionForm from './Form'
 
 export default function Caption(props) {
   const {
-    location: { search },
+    location: { search }, match: { url }
   } = props
   const token = new URLSearchParams(decodeURI(search)).get('token')
   if (token) {
     return <CaptionForm token={token} />
   } else {
-    return <Choose {...props} />
+    return <Choose {...props} next={url} caption />
   }
 }
