+++ src/token/choose/Search.jsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import seachStyle from './Search.module.scss'
-import inputStyle from '../component/input.module.scss'
+import inputStyle from '../../component/input.module.scss'
 import searchIcon from './search.svg'
 import { useTranslation } from 'react-i18next'
-import useStyle from '../component/useStyle'
+import useStyle from '../../component/useStyle'
 
 export default function Search({ searchTxt, setSearchTxt }) {
   const { t } = useTranslation()
